+++ rust/src/api/authentication.rs
@@ -1,5 +1,24 @@
-use super::PostOptions;
 use crate::{error::Result, models::*, Configuration};
+
+#[derive(Default)]
+pub struct AuthenticationAppPortalAccessOptions {
+    pub idempotency_key: Option<String>,
+}
+
+#[derive(Default)]
+pub struct AuthenticationExpireAllOptions {
+    pub idempotency_key: Option<String>,
+}
+
+#[derive(Default)]
+pub struct AuthenticationDashboardAccessOptions {
+    pub idempotency_key: Option<String>,
+}
+
+#[derive(Default)]
+pub struct AuthenticationLogoutOptions {
+    pub idempotency_key: Option<String>,
+}
 
 pub struct Authentication<'a> {
     cfg: &'a Configuration,
@@ -16,17 +35,17 @@ impl<'a> Authentication<'a> {
         &self,
         app_id: String,
         app_portal_access_in: AppPortalAccessIn,
-        options: Option<PostOptions>,
+        options: Option<AuthenticationAppPortalAccessOptions>,
     ) -> Result<AppPortalAccessOut> {
-        let PostOptions { idempotency_key } = options.unwrap_or_default();
+        let AuthenticationAppPortalAccessOptions { idempotency_key } = options.unwrap_or_default();
 
         crate::request::Request::new(
             http1::Method::POST,
             "/api/v1/auth/app-portal-access/{app_id}",
         )
         .with_path_param("app_id", app_id)
-        .with_body_param(app_portal_access_in)
         .with_optional_header_param("idempotency-key", idempotency_key)
+        .with_body_param(app_portal_access_in)
         .execute(self.cfg)
         .await
     }
@@ -36,14 +55,14 @@ impl<'a> Authentication<'a> {
         &self,
         app_id: String,
         application_token_expire_in: ApplicationTokenExpireIn,
-        options: Option<PostOptions>,
+        options: Option<AuthenticationExpireAllOptions>,
     ) -> Result<()> {
-        let PostOptions { idempotency_key } = options.unwrap_or_default();
+        let AuthenticationExpireAllOptions { idempotency_key } = options.unwrap_or_default();
 
         crate::request::Request::new(http1::Method::POST, "/api/v1/auth/app/{app_id}/expire-all")
             .with_path_param("app_id", app_id)
-            .with_body_param(application_token_expire_in)
             .with_optional_header_param("idempotency-key", idempotency_key)
+            .with_body_param(application_token_expire_in)
             .returns_nothing()
             .execute(self.cfg)
             .await
@@ -57,9 +76,9 @@ impl<'a> Authentication<'a> {
     pub async fn dashboard_access(
         &self,
         app_id: String,
-        options: Option<PostOptions>,
+        options: Option<AuthenticationDashboardAccessOptions>,
     ) -> Result<DashboardAccessOut> {
-        let PostOptions { idempotency_key } = options.unwrap_or_default();
+        let AuthenticationDashboardAccessOptions { idempotency_key } = options.unwrap_or_default();
 
         crate::request::Request::new(
             http1::Method::POST,
@@ -74,8 +93,8 @@ impl<'a> Authentication<'a> {
     /// Logout an app token.
     ///
     /// Trying to log out other tokens will fail.
-    pub async fn logout(&self, options: Option<PostOptions>) -> Result<()> {
-        let PostOptions { idempotency_key } = options.unwrap_or_default();
+    pub async fn logout(&self, options: Option<AuthenticationLogoutOptions>) -> Result<()> {
+        let AuthenticationLogoutOptions { idempotency_key } = options.unwrap_or_default();
 
         crate::request::Request::new(http1::Method::POST, "/api/v1/auth/logout")
             .with_optional_header_param("idempotency-key", idempotency_key)